LayoutTests/imported/w3c/web-platform-tests/content-security-policy/wasm-unsafe-eval/script-src-unsafe-eval-allows-wasm.any.sharedworker-expected.txt


PASS WebAssembly.instantiate() is allowed
PASS WebAssembly.compile() is allowed
PASS new WebAssembly.Module() is allowed
PASS WebAssembly.compileStreaming() is allowed
PASS WebAssembly.instantiateStreaming() is allowed
